+++ editor.planx.uk/src/pages/FlowEditor/components/FeedbackLog/FeedbackLog.stories.tsx
@@ -2,7 +2,7 @@ import Box from "@mui/material/Box";
 import { Meta, StoryObj } from "@storybook/react";
 import React from "react";
 
-import { FeedbackLog } from "../../FeedbackLog/FeedbackLog";
+import { FeedbackLog } from "./FeedbackLog";
 import { mockFeedback } from "./mocks/mockFeedback";
 
 const meta = {
