+++ editor.planx.uk/src/pages/FlowEditor/components/FeedbackLog/FeedbackLog.tsx
@@ -1,8 +1,6 @@
 import Typography from "@mui/material/Typography";
 import { GridFilterItem } from "@mui/x-data-grid";
 import { format } from "date-fns";
-import gql from "graphql-tag";
-import { client } from "lib/graphql";
 import capitalize from "lodash/capitalize";
 import React from "react";
 import { Feedback } from "routes/feedback";
@@ -21,28 +19,13 @@ import ErrorSummary from "ui/shared/ErrorSummary/ErrorSummary";
 import { ExpandableHelpText } from "./components/ExpandableHelpText";
 import { StatusChip } from "./components/StatusChip";
 import { feedbackTypeOptions, statusOptions } from "./feedbackFilterOptions";
+import { updateEditorNotes } from "./queries/updateEditorNotes";
 import { FeedbackLogProps } from "./types";
 import { EmojiRating, feedbackTypeText, stripHTMLTags } from "./utils";
 
 export const FeedbackLog: React.FC<FeedbackLogProps> = ({ feedback }) => {
   const handleProcessRowUpdate = async (updatedRow: Feedback) => {
-    await client.mutate({
-      mutation: gql`
-        mutation UpdateEditorNotes($id: Int!, $editor_notes: String) {
-          update_feedback_by_pk(
-            pk_columns: { id: $id }
-            _set: { editor_notes: $editor_notes }
-          ) {
-            id
-          }
-        }
-      `,
-      variables: {
-        id: updatedRow.id,
-        editor_notes: updatedRow.editorNotes,
-      },
-    });
-
+    await updateEditorNotes(updatedRow);
     return updatedRow;
   };
